# openapi.yaml
openapi: 3.1.0
info:
  title: Aave Liquidation Protection API
  version: 1.0.0
  description: |
    Autonomous Aave liquidation protection service powered by the Reactive Network.

    Agents can use this API to protect DeFi positions from liquidation on Aave (Base Sepolia).
    The service monitors health factors every ~12 minutes and automatically supplies collateral
    or repays debt when the health factor drops below a configured threshold.

    ## How it works
    1. Call `GET /api/services` to see available services and pricing.
    2. Call `POST /api/quote` to get an exact price for your desired duration.
    3. Call `POST /api/protect/liquidation` with an x402 payment header to activate protection.
       - The endpoint is gated by the [x402 payment protocol](https://x402.org).
       - If no payment is included, the server returns `402 Payment Required` with payment details.
       - Payment is in USDC on Base Sepolia (eip155:84532).
    4. After payment, approve the `AaveProtectionCallback` contract (returned in the response)
       to spend your collateral and/or debt assets.
    5. Manage the config with pause/resume/cancel. Monitor it via the status endpoints.

    ## x402 Payment Flow
    Protected endpoints return `402 Payment Required` when called without payment.
    The response body contains `accepts` — an array of payment options. Agents must:
    1. Sign an EIP-3009 `transferWithAuthorization` over USDC.
...
    3. Retry the request.

...

    ## Networks
...

...
servers:
...
paths:
...
  /api/quote:
    post:
      operationId: getQuote
      tags: [Discovery]
      summary: Get exact price quote for a service
      description: |
        Returns the exact USDC price (in base units, 6 decimals) for a given service and duration.
        Price includes a 20% gas buffer on top of the base rate.

        Base rate: $0.25/day for `aave-protection`.

        Use this before calling the protected endpoint so you know exactly how much USDC to authorize.
        No payment required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequest"
            examples:
              oneDay:
                summary: Quote for 1 day
                value:
                  service: aave-protection
                  durationSeconds: 86400
              sevenDays:
                summary: Quote for 7 days
                value:
                  service: aave-protection
                  durationSeconds: 604800
      responses:
        "200":
          description: Price quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "400":
          description: Invalid service ID or duration out of range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/protect/liquidation:
    post:
      operationId: createLiquidationProtection
      tags: [Protection]
      summary: Create Aave liquidation protection (x402-gated)
      description: |
        Creates an on-chain protection configuration. The Reactive Contract will then monitor
        the protected user's Aave health factor every ~12 minutes and execute protection
        (collateral deposit, debt repayment, or both) when it drops below the threshold.

        **This endpoint requires x402 payment in USDC on Base Sepolia (eip155:84532).**

        ### Without payment
        Returns `402 Payment Required` with an `accepts` array describing how to pay.
        The price is dynamic — computed from the `duration` field in the request body.

        ### With payment
        Attach the signed EIP-3009 authorization as the `X-PAYMENT` header (base64-encoded JSON).
        On success, returns the `configId` and the AaveProtectionCallback address that needs
        token approval before protection can execute.

        ### After creation
        You MUST approve the `AaveProtectionCallback` contract to spend:
        - `collateralAsset` tokens (for collateral deposit mode)
        - `debtAsset` tokens (for debt repayment mode)
        Both if `protectionType` is `2` (BOTH).

        ### Protection types
        | Value | Name | Description |
        |-------|------|-------------|
        | 0 | COLLATERAL_DEPOSIT | Supplies collateral to Aave to raise HF |
        | 1 | DEBT_REPAYMENT | Repays debt to raise HF |
        | 2 | BOTH | Tries collateral first, falls back to debt repayment |

        ### Health factor encoding
        Health factors are WAD-encoded (18 decimals). Examples:
        - `1.5 HF` → `"1500000000000000000"`
        - `2.0 HF` → `"2000000000000000000"`
      security:
        - x402: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProtectionRequest"
            examples:
              defaultProtection:
                summary: Protect with collateral deposit, 1 day
                value:
                  protectedUser: "0xYourWalletAddress"
                  protectionType: 0
                  healthFactorThreshold: "1500000000000000000"
                  targetHealthFactor: "2000000000000000000"
                  collateralAsset: "0x4200000000000000000000000000000000000006"
                  debtAsset: "0xba50Cd2A20f6DA35D788639E581bca8d0B5d4D5f"
                  preferDebtRepayment: false
                  duration: 86400
      responses:
        "200":
          description: Protection config created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProtectionResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "402":
          description: Payment required — attach x402 payment header and retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequiredResponse"
        "500":
          description: On-chain transaction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Reactive Contract underfunded — service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/protect/liquidation/pause:
    post:
      operationId: pauseLiquidationProtection
      tags: [Protection]
      summary: Pause a protection config
      description: |
        Pauses an active protection config. While paused, the Reactive Contract will skip
        health factor checks for this config. No payment required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigIdRequest"
      responses:
        "200":
          description: Config paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigActionResponse"
        "400":
          description: Invalid configId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: On-chain transaction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/protect/liquidation/resume:
    post:
      operationId: resumeLiquidationProtection
      tags: [Protection]
      summary: Resume a paused protection config
      description: |
        Resumes a previously paused protection config. Health factor monitoring will
        resume on the next Reactive callback cycle (~12 minutes). No payment required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigIdRequest"
      responses:
        "200":
          description: Config resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigActionResponse"
        "400":
          description: Invalid configId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: On-chain transaction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/protect/liquidation/cancel:
    post:
      operationId: cancelLiquidationProtection
      tags: [Protection]
      summary: Cancel a protection config permanently
      description: |
        Permanently cancels a protection config. This action is irreversible — a cancelled
        config cannot be resumed. No payment required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigIdRequest"
      responses:
        "200":
          description: Config cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigActionResponse"
        "400":
          description: Invalid configId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: On-chain transaction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/status/config/{configId}:
    get:
      operationId: getProtectionConfig
      tags: [Status]
      summary: Get protection config details
      description: |
        Fetches the on-chain state of a protection config including its current status,
        execution history, and all configuration parameters. No payment required.
      parameters:
        - name: configId
          in: path
          required: true
          description: The numeric config ID returned when the config was created
          schema:
            type: integer
            minimum: 0
          example: 42
      responses:
        "200":
          description: Config details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtectionConfigResponse"
        "400":
          description: Invalid config ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to fetch from chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/approve/permit:
    post:
      operationId: relayPermit
      tags: [Protection]
      summary: Relay an EIP-2612 permit signature on-chain (free)
      description: |
        HTTP-only agents (e.g. OpenClaw-style) can pay via x402 but cannot submit EVM transactions
        to approve the callback contract to spend their tokens. This endpoint accepts a pre-signed
        EIP-2612 permit and relays it on-chain — the server pays the Base Sepolia gas.

        **Only works for EIP-2612 tokens (USDC). WETH does not support permit.**

        This endpoint is free — no x402 payment needed. The fee is already included in the
        protection service price.

        ### Typical HTTP-only agent flow
        1. `GET /api/services` — discover service and CC address
        2. `POST /api/quote` — get price for duration
        3. Sign EIP-2612 `permit(owner, CC_address, amount, deadline)` off-chain
        4. `POST /api/approve/permit` — relay the permit (this endpoint)
        5. `POST /api/protect/liquidation` (with x402 payment) — register protection

        ### Limitation
        HTTP-only agents can only use `protectionType: 1` (DEBT_REPAYMENT with Aave USDC).
        `protectionType: 0` (COLLATERAL_DEPOSIT with WETH) requires a direct EVM approval
        transaction because WETH has no permit support.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermitRequest"
            examples:
              usdcPermit:
                summary: Permit Aave USDC spend for debt repayment
                value:
                  token: "0xba50Cd2A20f6DA35D788639E581bca8d0B5d4D5f"
                  owner: "0xYourWalletAddress"
                  spender: "0xAaveProtectionCallbackAddress"
                  value: "1000000000"
                  deadline: 1712000000
                  v: 27
                  r: "0xabc123..."
                  s: "0xdef456..."
      responses:
        "200":
          description: Permit relayed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermitResponse"
        "400":
          description: Invalid parameters or WETH address provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: On-chain permit transaction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/status/configs:
    get:
      operationId: listActiveConfigs
      tags: [Status]
      summary: List all active protection config IDs
      description: |
        Returns all currently active (non-cancelled, non-paused) protection config IDs
        from the on-chain contract. No payment required.
      responses:
        "200":
          description: List of active config IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveConfigsResponse"
        "500":
          description: Failed to fetch from chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    x402:
      type: apiKey
      in: header
      name: X-PAYMENT
      description: |
        Base64-encoded JSON containing a signed EIP-3009 `transferWithAuthorization`
        over USDC on Base Sepolia. Signed locally by the agent's private key using the
        x402 SDK — no separate facilitator call needed by the client. The facilitator
        is called server-side to verify and settle the payment.
        See https://x402.org/docs for SDK helpers (`@x402/fetch`, `@x402/evm`).
  schemas:
...
    QuoteRequest:
      type: object
      required: [service, durationSeconds]
      properties:
        service:
          type: string
          description: Service ID from the catalog
          enum: [aave-protection]
          example: aave-protection
        durationSeconds:
          type: integer
          description: Desired protection duration in seconds (min 3600, max 2592000)
          minimum: 3600
          maximum: 2592000
          example: 86400
...
    ConfigIdRequest:
      type: object
      required: [configId]
      properties:
        configId:
          type: integer
          minimum: 0
          description: Numeric protection config ID
          example: 42

    # ── Response schemas ─────────────────────────────────────────────────────────
...
    QuoteResponse:
      type: object
      properties:
        service:
          type: string
          example: aave-protection
        durationSeconds:
          type: integer
          example: 86400
        price:
          type: string
          description: Human-readable total price
          example: "$0.3"
        priceBaseUnits:
          type: string
          description: Total price in USDC base units (6 decimals) — use this for EIP-3009 authorization
          example: "300000"
        currency:
          type: string
          example: USDC
        network:
          type: string
          example: eip155:84532
...
    ConfigActionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        configId:
          type: integer
          example: 42
        txHash:
          type: string
          description: Base Sepolia transaction hash
          example: "0xabc123..."
        action:
          type: string
          enum: [paused, resumed, cancelled]
          example: paused
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "On-chain config creation failed"
        reason:
          type: string
          description: Detailed reason (often from the on-chain revert)
          example: "execution reverted: insufficient balance"
...
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid parameters"
        details:
          type: object
          description: Per-field validation errors
          additionalProperties:
            type: array
            items:
              type: string
          example:
            protectedUser: ["Invalid address format"]
            protectionType: ["Expected number, received string"]